Set `DYLD_LIBRARY_PATH` to run Swift tests instead of rpath
When running the Swift tests, I realized it's better to set
`DYLD_LIBRARY_PATH` in the environment to find
`libmyanmartools.dylib`, rather than hard-coding rpath to the absolute
path of the build directory.

Also, I forgot to quote $(CURDIR) so it would fail if the current
directory had a space in it.